+++ src/containerapp/HISTORY.rst
@@ -7,6 +7,7 @@ Upcoming
 ++++++
 * 'az containerapp job execution show/list': improve table output format
 * 'az containerapp create/update': --yaml support properties for api-version 2023-04-01-preview (e.g. subPath, mountOptions)
+* 'az containerapp service': add support for creation and deletion of kafka
 
 0.3.33
 ++++++